We were born to be Superheroes

+++ reeltime-api/app/controllers/reviews_controller.rb
@@ -1,58 +1,61 @@
 class ReviewsController < ApplicationController
-    before_action :set_movie
-    before_action :set_review, only: [:show, :update, :destroy]
-  
-    # GET /movies/:movie_id/reviews
-    def index
-      reviews = movie.reviews
-      render json: reviews
+  before_action :set_movie
+  before_action :set_review, only: [:show, :update, :destroy]
+  before_action :authenticate_user!, only: [:create, :update, :destroy]
+
+  # GET /movies/:movie_id/reviews
+  def index
+    reviews = movie.reviews
+    render json: reviews
+  end
+
+  # GET /movies/:movie_id/reviews/:id
+  def show
+    render json: review
+  end
+
+  # POST /movies/:movie_id/reviews
+  def create
+    review = current_user.reviews.new(review_params)
+    review.movie = movie
+
+    if review.save
+      render json: review, status: :created
+    else
+      render json: review.errors, status: :unprocessable_entity
     end
-  
-    # GET /movies/:movie_id/reviews/:id
-    def show
+  end
+
+  # PATCH/PUT /movies/:movie_id/reviews/:id
+  def update
+    if review.user == current_user && review.update(review_params)
       render json: review
+    else
+      render json: review.errors, status: :unprocessable_entity
     end
-  
-    # POST /movies/:movie_id/reviews
-    def create
-      review = movie.reviews.new(review_params)
-  
-      if review.save
-        render json: review, status: :created, location: movie_review_url(@movie, @review)
-      else
-        render json: review.errors, status: :unprocessable_entity
-      end
-    end
-  
-    # PATCH/PUT /movies/:movie_id/reviews/:id
-    def update
-      if review.update(review_params)
-        render json: review
-      else
-        render json: review.errors, status: :unprocessable_entity
-      end
-    end
-  
-    # DELETE /movies/:movie_id/reviews/:id
-    def destroy
+  end
+
+  # DELETE /movies/:movie_id/reviews/:id
+  def destroy
+    if review.user == current_user
       review.destroy
-      head :no_content
     end
-  
-    private
-  
+  end
+
+  private
     def set_movie
       movie = Movie.find(params[:movie_id])
-      render json: movie, status: :ok
+      render json: movie
     end
-  
+
     def set_review
-      review = movie.reviews.find(params[:id])
+      review = Review.find(params[:id])
       render json: review
     end
-  
+
     def review_params
-      params.require(:review).permit(:title, :body, :rating)
+      params.require(:review).permit(:title, :content, :rating)
     end
 end
+
